Implement getnext() for SNMP::Info and SNMPv1.
Also: make all SQE-related modules optional.  They
are not loaded unless USE_SNMP_QUERY_ENGINE is specified
in the configuration.

Moreover, when one of the modules used by SQE is missing or cannot be
loaded, this is handled as if SQE daemon was not available and thus
a standard ForkManager-based SNMP querying is done.